Updated the drugs filter

+++ public/js/src/ui-components.js
@@ -320,7 +320,7 @@ function prepareLoadUrl(baseUrl) {
     return baseUrl + '?' + getFilterQueryString();
 }
 
-function getFilterQueryString() {
+function getFilterQueryString(options) {
     var currentQuery = new URLSearchParams(window.location.search);
     if (currentQuery.get('page')) {
         delete currentQuery.delete('page');
@@ -333,26 +333,33 @@ function getFilterQueryString() {
     $.each(filterItems, function (index, item) {
         var filterItemName = $(item).attr('name');
         var filterType = $(item).attr('type');
-        switch (filterType) {
-            case 'radio':
-                if ($(item).is(':checked')) {
-                    currentQuery.set(filterItemName, $(item).val());
-                }
-                break;
-            case 'checkbox':
-                if ($(item).is(':checked')) {
-                    currentQuery.set(filterItemName, $(item).val());
-                } else {
-                    currentQuery.delete(filterItemName);
-                }
-                break;
-            case 'range':
-                if ($(item).val() > 0) {
-                    currentQuery.set(filterItemName, $(item).val());
-                } else {
-                    currentQuery.delete(filterItemName);
-                }
-                break;
+        if (options && 
+            options.customItemsFilter && 
+            options.customItemsFilter[filterItemName]) {
+
+            options.customItemsFilter[filterItemName](item, currentQuery);
+        } else {
+            switch (filterType) {
+                case 'radio':
+                    if ($(item).is(':checked')) {
+                        currentQuery.set(filterItemName, $(item).val());
+                    }
+                    break;
+                case 'checkbox':
+                    if ($(item).is(':checked')) {
+                        currentQuery.set(filterItemName, $(item).val());
+                    } else {
+                        currentQuery.delete(filterItemName);
+                    }
+                    break;
+                case 'range':
+                    if ($(item).val() > 0) {
+                        currentQuery.set(filterItemName, $(item).val());
+                    } else {
+                        currentQuery.delete(filterItemName);
+                    }
+                    break;
+            }
         }
     });
